ASC-78 Moved each instrument python class to a folder called instruments, moved super class Instrument into atlast_sc
